Move telegram_bot services to dedicated services module
