+++ tests/components/telegram_bot/test_telegram_bot.py
@@ -26,7 +26,6 @@ from telegram.error import (
     TimedOut,
 )
 
-from homeassistant.components.telegram_bot import ATTR_LATITUDE, ATTR_LONGITUDE
 from homeassistant.components.telegram_bot.bot import ALLOWED_UPDATES
 from homeassistant.components.telegram_bot.const import (
     ATTR_AUTHENTICATION,
@@ -92,6 +91,8 @@ from homeassistant.config_entries import ConfigEntryState
 from homeassistant.const import (
     ATTR_DOMAIN,
     ATTR_ENTITY_ID,
+    ATTR_LATITUDE,
+    ATTR_LONGITUDE,
     ATTR_SERVICE,
     CONF_API_KEY,
     CONF_PLATFORM,
